requirements.txt
boto3==1.9.189
auth0-python==3.9.1
pandas==0.24.2